requirements.txt
Quart~=0.19.3
pyjwt~=2.8.0
aiohttp~=3.8.6
oauthlib~=3.2.2
discord.py~=2.3.2
cachetools~=5.3.2
setuptools~=68.2.0
Async-OAuthlib~=0.0.9
py-bcrypt~=0.4